requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
